Add a feature for interactive mode
A chain of nested functions has been created. During the operation of
each of them, the name of the variable is displayed in the console,
with a suggestion to enter a value. Next is a check on the correctness
of the format. If it is incorrect, an error is displayed in the console.
If it is correct, the same function is called for the next variable.
In the end the function of calculation of roots is called.