Add a non-interactive program feature
Added a variable equal to the value of the parameter passed at startup.
This should be the path to the variable value file. The function reads
the contents of the file, if an error occurs, the program terminates.
If not, the values are reduced to a number format. Variable a is checked
for a non-zero value, if it is equal to 0, an error message is displayed
and the program terminates. Then the variables are checked for the type
of number and the function of calculating the roots is called. Added
text.txt file for verification.